chore(store): renamed store files
