+++ src/utils/firebase/firebase.ts
@@ -1,11 +1,13 @@
 import firebase from "firebase/app";
 
-if (!process.env.REACT_APP_FB_KEY) {
+import { FB_KEY } from "../environmentVariables";
+
+if (!FB_KEY) {
   console.warn(`No database key set.`);
 }
 
 firebase.initializeApp({
-  apiKey: process.env.REACT_APP_FB_KEY,
+  apiKey: FB_KEY,
   authDomain: window.location.hostname,
   projectId: "wbtc-portal",
 });
